APPSERV-149 Move local instance command under original command
